Use residual battery value Showing parameter bounds Start to add network logging store

+++ app/src/main/java/com/alpriest/energystats/models/BatteryResponse.kt
@@ -3,7 +3,7 @@ package com.alpriest.energystats.models
 data class BatteryResponse(
     val power: Double,
     val soc: Int,
-    val residual: Double,
+    val residual: Int,
     val temperature: Double
 )
